Update lily58.keymap file
The lily58.keymap file has been updated to change the key binding for the left shift key. Previously, it was bound to the LSHIFT key, but now it is bound to the LGUI key. This change was made to improve the usability and functionality of the keyboard layout. Additionally, the sensor bindings for volume control have been modified to use the inc_dec_kp sensor.